Rename WC module package 'main' to 'session'
