+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/session/WalletConnectSessionViewModel.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.walletconnect.main
+package io.horizontalsystems.bankwallet.modules.walletconnect.session
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
